:green_heart: split tests into separated files
